Fix issue with logging
