+++ connector-packager/connector_packager/jar_jdk_signer.py
@@ -12,7 +12,7 @@ from .helper import check_jdk_environ_variable
 JARSIGNER_EXECUTABLE_NAME = "jarsigner"
 if os.name == 'nt':
     JARSIGNER_EXECUTABLE_NAME += ".exe"
-logger = logging.getLogger(__name__)
+logger = logging.getLogger('packager_logger')
 
 KEYSTORE_PWD_PROMPT_LENGTH = len("Enter Passphrase for keystore: ")
 ALIAS_PWD_PROMPT_LENGTH = len("Enter key password for : ")
